Wrote some documenting files, mostly aspirations and milestones.
Further a lot of stuff:

The macros of struct are now use an extension via a plist of the state. This makes it more library-like.

Wrote a unit test for type-get-var and type-fill. Getting type data into macros seems with those seems to be working correctly. Moved eql and any to
their own files in types/ but i can't really consider them more library-like. (Mac-get plays with them directly through type-get-var.)

Got do-times, to work, it required: (It is cool that do-times could have been defined in a library, if lang could run macros.)
* In types/numbers.lisp i added eql-types to the manual-generality function.
* Just evaluating a number used to make a eql-type, now it produces the default integer type. (TODO need to choose higher as needed by the number)
  (const whatever) now produces the eql-type. (Might need to just check whether it can be an eql type, or retroactively change it once the resolver
finds out that it is not an eql-type.)
I also changed the out-type of +,-,*/ operators so that small integer/float types has the default type as out-type.

Added into progn:
* removing progns within.
* removing itself if only one argument. Fixed to-c.lisp so that it accepts that.
Also let now detects if there are zero variables in a let.

The unit test of typeset now seems to have misses. ~1 a 2000. Need to fix that.