Improved website, now urls are representative for links. (Frames seem stupid for not providing the feature itself.) Some stuff is combined with a little shell script now.
Got pointers and references working afaics and afaik. It does not work for the conversion to C yet.(Doesn't recognize it as special type i think.)

Removed some of the messing with the resolved code in to-c.lisp. Renamed manual-type-generality to manual-type-coarser, and named each of them.
(Probably need to improve that code though.) Renamed some files not to have part of their name repeated in the directory.