Added a simple macro expander, and an nested list to xml converter. (Bit on the side of the project)
Separated some of the namespace code into its own package, will have to see if it makes sense to also do the rest of it.
Put the argumentizer (argument-positioner.lisp) into the generic package.

Some modifications to resolve are to make namespaces work better; the values didn't give the correct variable name.

Messed about with typeset a little, still have not figured out how to make converters work properly.. Ref is turning out not to be that easy, at
least. (Maybe i must manual it..) The type coarseness should now have access to variables encountered, and manual coarseness gets that as an
additional argument.