Removed type-coarseness based on conversion functions; have to figure out how to make that work properly.
Added a way to allow for variables to be used as references into functions specificly. A todo is whether to have that as a requirement for setting in
general. (Or have a const in the same way?) Another thing todo is references for settable-functions.

Replaced play-test-util.lisp with chain.lisp. This file contains functions and a macro that chains the operations of: reading from string, stream,
file; resolving code; summarising resolved code, and converting it to C.