Made a new reader (again) this one should be able to read C code. It now also reads the lisp code, there is code to support XML-like listing, but i haven't tested it yet. (Reader as a whole is not unit tested.)
I started working on proofs and definitions for the type system. It is not quite complete yet. Proving that the thing with variables matching still
is ordered might be tricky. Also don't know about modifications yet. Formalising might help in finding out how to use conversion for type coarseness,
and improving the typeset and coarseness code as a whole.

Also made namespace package keep more to itself, but i don't think i will be able to avoid using write-namespace and namespaces. Removed
type-eql-like, as it was silly, might need to check if it actually implies equality.

Added a macro for adding functions that are created at converting time. Messed around a little trying to get strings.