Made new to-c.lisp, old one will be removed next. This splits up the treatment of the different macros and does it a little differently too.
Put argumentizing in its own package, and improved it with a new function that parses arguments, which is then used for argumentize-list.

Made chain.lisp more general, making it a thing that can convert between arbitrary things, if the conversion functions are properly defined.
(Should've looked for a library for this..)

Apparently umacs; a thing that did something like iterate; was in committed in here too, it isn't anymore now. I expect that i will start using
iterate instead of loop soon.