HUGE AMMOUNT OF CHANGES
I shouldnt have waited so long to commit. It is still a little too messy imo.

New approach, types are now based on either resolvation or overridden by functions that determine them. CLOS like things is now expected to be
possible with a library. (Or you code the choosing of the types and it is optimized away.)

This meant a whole new approach, don't know where to start really:
* The majorly different approach, deleted all the type-coarseness/utility functions. Not that the code was that bad, other approach does not need
it. The library to do the CLOS type things might have use for them again.
* Instead of an assoclist for variables it now has a local state for local variables, macros, symbol-macros and functions.
* Resolved output is now completely build from non-lists. No more raw-macros they're all 'raw'macros. (making nonraw macros matter of making a macro
and restricting access to resolving.)

Note that the conversion to C entirely doesn't work.(Shouldnt be hard to get it to work again.)